update privacy api
